move usingDataSourceExecWithIncompatTypes* to CometTestBase
