+++ spark/src/test/scala/org/apache/comet/exec/CometColumnarShuffleSuite.scala
@@ -758,7 +758,7 @@ abstract class CometColumnarShuffleSuite extends CometTestBase with AdaptiveSpar
         // TODO: revisit this when we have resolution of https://github.com/apache/arrow-rs/issues/7040
         // and https://github.com/apache/arrow-rs/issues/7097
         val fieldsToTest =
-          if (CometSparkSessionExtensions.usingDataSourceExec(conf)) {
+          if (usingDataSourceExec(conf)) {
             Seq(
               $"_1",
               $"_4",
